+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Next Release
 
+<!-- insertion marker -->
+## [0.3.0](https://github.com/repo/owner/releases/tag/0.3.0) - 2026-09-22
+
 ### Features
 
 #### CPP Rules
@@ -16,7 +19,6 @@ All notable changes to this project will be documented in this file.
 
 - Invalidate the incremental state file whenever the TOML config file changes (a hash of its contents is stored in the state file), so all files are re-checked with the new configuration
 
-<!-- insertion marker -->
 ## [0.3.0](https://github.com/repo/owner/releases/tag/0.3.0) - 2026-09-20
 
 ### Features
@@ -185,6 +187,7 @@ All notable changes to this project will be documented in this file.
 - Add automatic release CI for PRs to main (either via title or via hotfix/ branch)
 - Add overnight CI runs for pytest and ruff CIs
 - Add test coverage to pytest CI
+